apps/client_backend/pyproject.toml
[tool.poetry]
name = "client-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
python-dotenv = "^1.0.0"
pydub = "^0.25.1"
nltk = "^3.8.1"
openai = "^0.27.8"
sounddevice = "^0.4.6"
azure-cognitiveservices-speech = "^1.27.0"
websockets = "^11.0.2"
requests = "^2.28.2"
uvicorn = "^0.21.1"
fastapi = "^0.100.1"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
six = "^1.16.0"
opentelemetry-sdk = "^1.17.0"
janus = "^1.0.0"
scipy = "^1.10.1"
anthropic = "^0.2.9"
elevenlabs = "^0.2.6"
google-cloud-texttospeech = "^2.14.1"
gtts = { version = "^2.3.1", optional = true}
google-cloud-speech = "^2.19.0"
redis = "^4.5.4"
twilio = "^8.1.0"
vonage = { version = "^3.5.1", optional = true}
nylas = "^5.14.0"
speechrecognition = "^3.10.0"
aiohttp = "^3.8.4"
langchain = "^0.0.198"
google-cloud-aiplatform = {version = "^1.26.0", optional = true}
miniaudio = "^1.59"
boto3 = "^1.28.28"

[tool.poetry.extras]
synthesizers = ["gtts", "google-cloud-texttospeech", "elevenlabs"]
transcribers = ["google-cloud-speech"]
telephony = ["twilio", "redis", "vonage"]
agents = ["google-cloud-aiplatform"]
actions = ["nylas"]
all = ["gtts", "google-cloud-texttospeech", "elevenlabs", "google-cloud-speech", "google-cloud-aiplatform", "twilio", "redis", "nylas", "vonage"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
